Changing: command -> request command_result -> response ConnectionHandler -> ConnHandler ConnectionHandlers -> ConnSet
Fixing some bugs.

+++ Server/SocketConnection.cpp
@@ -41,7 +41,7 @@ int SocketConnection::send_buffer(char *buffer, size_t buf_size) {
 int SocketConnection::recv_data(char *recvbuf, size_t max_bufsize) {
     int result;
     int read_bytes = 0;
-    uint16_t buf_size = 0;
+    size_t buf_size = 0;
 
     result = recv(conn_socket, (char *)&buf_size, sizeof(buf_size), 0);
     if (!check_connection_result(result)) {
@@ -51,7 +51,7 @@ int SocketConnection::recv_data(char *recvbuf, size_t max_bufsize) {
     if ((result != sizeof(buf_size)) || (buf_size > max_bufsize)) {
         throw "invalid message";
     }
-// "Should receive " << buf_size << " bytes" << endl;
+//    cout << "Should receive " << buf_size << " bytes" << endl;
 
     // Receive until the peer closes the connection
     while(buf_size > read_bytes) {
